フォルダの整頓をしました、名前の修正や余分なストラクトを減らしました。
